Relocate numba specific tests in test_diagnostics (and skip if not installed)
